Fix checked graph bug
Previously, the radio button for which Network Graph is being displayed
would always reset to the first choice after click even if the screen
was rendering the second graph. Fix this by creating a new state
variable where the radio button can evaluate whether or not it should be
checked according to this state's value.